Move backendStatus action creators to separate folder
