+++ src/pages/data-management/index.jsx
@@ -17,7 +17,7 @@ import ProjectDetails from '../../components/data-management/ProjectDetails';
 import LoadingModal from '../../components/LoadingModal';
 
 import { loadProcessingSettings } from '../../redux/actions/experimentSettings';
-import loadBackendStatus from '../../redux/actions/experimentSettings/backendStatus/loadBackendStatus';
+import loadBackendStatus from '../../redux/actions/backendStatus/loadBackendStatus';
 
 const DataManagementPage = ({ route }) => {
   const dispatch = useDispatch();
