+++ src/redux/actions/backendStatus/index.js
@@ -1,0 +1,7 @@
+import loadBackendStatus from './loadBackendStatus';
+import updateBackendStatus from './updateBackendStatus';
+
+export {
+  loadBackendStatus,
+  updateBackendStatus,
+};
